Add path tests
