+++ src/parser/path.rs
@@ -36,6 +36,21 @@ mod tests {
     use super::*;
 
     #[test]
+    fn test_parse_path() {
+        let data: Vec<(&str, Path)> = vec![
+            ("003@.0", Path::new("003@", None, '0', None)),
+            ("  003@.0 ", Path::new("003@", None, '0', None)),
+            ("012A/01.0", Path::new("012A", Some("01"), '0', None)),
+            ("003@.0[0]", Path::new("003@", None, '0', Some(0))),
+            ("012A/00.0[1]", Path::new("012A", Some("00"), '0', Some(1))),
+        ];
+
+        for (input, expected) in data {
+            assert_eq!(parse_path(input), Ok(("", expected)));
+        }
+    }
+
+    #[test]
     fn test_parse_path_list() {
         let path_list =
             vec![Path::from_str("003@.0"), Path::from_str("002@.0[0]")];
