+++ src/path.rs
@@ -90,44 +90,3 @@ impl fmt::Display for Path {
         Ok(())
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_parse_path() {
-        let data: Vec<(&str, Path)> = vec![
-            ("003@.0", Path::new("003@", None, '0', None)),
-            ("  003@.0 ", Path::new("003@", None, '0', None)),
-            ("012A/01.0", Path::new("012A", Some("01"), '0', None)),
-            ("003@.0[0]", Path::new("003@", None, '0', Some(0))),
-            ("012A/00.0[1]", Path::new("012A", Some("00"), '0', Some(1))),
-        ];
-
-        for (input, expected) in data {
-            assert_eq!(parse_path(input), Ok(("", expected)));
-        }
-    }
-
-    #[test]
-    fn test_path_new() {
-        let path = Path::new("012A", Some("000"), '0', Some(0));
-        assert_eq!(path.tag(), "012A");
-        assert_eq!(path.occurrence(), Some("000"));
-        assert_eq!(path.code(), '0');
-        assert_eq!(path.index(), Some(0));
-    }
-
-    #[test]
-    fn test_path_from_str() {
-        let path = "012A/000.0".parse::<Path>().unwrap();
-        assert_eq!(path.tag(), "012A");
-        assert_eq!(path.occurrence(), Some("000"));
-        assert_eq!(path.code(), '0');
-        assert_eq!(path.index(), None);
-
-        let result = "003@.?".parse::<Path>();
-        assert_eq!(result.err(), Some(ParsePicaError::InvalidPath));
-    }
-}
